+++ sample/src/main/java/com/polidea/rxandroidble/sample/example4_characteristic/CharacteristicOperationExampleActivity.java
@@ -2,10 +2,12 @@ package com.polidea.rxandroidble.sample.example4_characteristic;
 
 import android.os.Bundle;
 import android.support.design.widget.Snackbar;
+import android.util.Log;
 import android.widget.Button;
 import android.widget.TextView;
 
 import com.polidea.rxandroidble.RxBleConnection;
+import com.polidea.rxandroidble.RxBleDevice;
 import com.polidea.rxandroidble.sample.DeviceActivity;
 import com.polidea.rxandroidble.sample.R;
 import com.polidea.rxandroidble.sample.SampleApplication;
@@ -26,7 +28,7 @@ import static com.trello.rxlifecycle.ActivityEvent.PAUSE;
 
 public class CharacteristicOperationExampleActivity extends RxAppCompatActivity {
 
-    public static String EXTRA_CHARACTERISTIC_UUID = "extra_uuid";
+    public static final String EXTRA_CHARACTERISTIC_UUID = "extra_uuid";
     @Bind(R.id.connect)
     Button connectButton;
     @Bind(R.id.read_output)
@@ -38,7 +40,7 @@ public class CharacteristicOperationExampleActivity extends RxAppCompatActivity 
     private UUID characteristicUuid;
     private PublishSubject<Void> disconnectTriggerSubject = PublishSubject.create();
     private Observable<RxBleConnection> connectionObservable;
-    private String macAddress;
+    private RxBleDevice bleDevice;
 
 
     @OnClick(R.id.read)
@@ -60,8 +62,15 @@ public class CharacteristicOperationExampleActivity extends RxAppCompatActivity 
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_example4);
         ButterKnife.bind(this);
-        macAddress = getIntent().getStringExtra(DeviceActivity.EXTRA_MAC_ADDRESS);
+        String macAddress = getIntent().getStringExtra(DeviceActivity.EXTRA_MAC_ADDRESS);
         characteristicUuid = (UUID) getIntent().getSerializableExtra(EXTRA_CHARACTERISTIC_UUID);
+        bleDevice = SampleApplication.getRxBleClient(this).getBleDevice(macAddress);
+        connectionObservable = bleDevice
+                .establishConnection(this, false)
+                .takeUntil(disconnectTriggerSubject)
+                .compose(bindUntilEvent(PAUSE))
+                .doOnUnsubscribe(this::clearSubscription)
+                .compose(new ConnectionSharingAdapter());
         //noinspection ConstantConditions
         getSupportActionBar().setSubtitle(getString(R.string.mac_address, macAddress));
     }
@@ -72,14 +81,8 @@ public class CharacteristicOperationExampleActivity extends RxAppCompatActivity 
         if (isConnected()) {
             triggerDisconnect();
         } else {
-            connectionObservable = SampleApplication.getRxBleClient(this)
-                    .getBleDevice(macAddress)
-                    .establishConnection(this, false)
-                    .takeUntil(disconnectTriggerSubject)
-                    .compose(bindUntilEvent(PAUSE))
-                    .doOnUnsubscribe(this::setNotConnected)
-                    .compose(new ConnectionSharingAdapter());
             connectionObservable.subscribe(rxBleConnection -> {
+                Log.d(getClass().getSimpleName(), "Hey, connection has been established!");
             }, this::onConnectionFailure);
         }
 
@@ -87,7 +90,7 @@ public class CharacteristicOperationExampleActivity extends RxAppCompatActivity 
     }
 
     private boolean isConnected() {
-        return connectionObservable != null;
+        return bleDevice.getConnectionState() == RxBleConnection.RxBleConnectionState.CONNECTED;
     }
 
     private void onConnectionFailure(Throwable throwable) {
@@ -100,7 +103,7 @@ public class CharacteristicOperationExampleActivity extends RxAppCompatActivity 
         Snackbar.make(findViewById(android.R.id.content), "Read error: " + throwable, Snackbar.LENGTH_SHORT).show();
     }
 
-    private void setNotConnected() {
+    private void clearSubscription() {
         connectionObservable = null;
         updateUI();
     }