+++ CHANGELOG.md
@@ -19,4 +19,7 @@ Version 0.0.3
 * Fixed error callback notifying about disconnects
 
 Version 0.0.4
+* Removed duplicated API for connection state from RxBleConnection
+* Renamed API for connection state observation in RxBleDevice
+* Added convenience method to check current connection state
 * Added ability to filter scan results with one service more easily